refactor(web): move usage analytics hooks into lib/usage
