+++ web/src/app/ee/admin/performance/usage/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { AdminDateRangeSelector } from "@/refresh-components/AdminDateRangeSelector";
-import { useTimeRange } from "@/app/ee/admin/performance/lib";
+import { useTimeRange } from "@/lib/usage/hooks";
 import PerUserUsagePanel from "@/views/admin/PerUserUsagePanel";
 import { ADMIN_ROUTES } from "@/lib/admin-routes";
 import { Divider } from "@opal/components";
