+++ web/src/lib/dateUtils.ts
@@ -28,6 +28,26 @@ export function getXDaysAgo(daysAgo: number) {
   return daysAgoDate;
 }
 
+export function convertDateToEndOfDay(date?: Date | null) {
+  if (!date) {
+    return date;
+  }
+
+  const dateCopy = new Date(date);
+  dateCopy.setHours(23, 59, 59, 999);
+  return dateCopy;
+}
+
+export function convertDateToStartOfDay(date?: Date | null) {
+  if (!date) {
+    return date;
+  }
+
+  const dateCopy = new Date(date);
+  dateCopy.setHours(0, 0, 0, 0);
+  return dateCopy;
+}
+
 export function getXYearsAgo(yearsAgo: number) {
   const today = new Date();
   const yearsAgoDate = new Date(today);
